did a bunch of stuff but keep getting 500 error.  don't think this will help but am at a loss

+++ src/main/java/com/example/demo/models/Adder.java
@@ -4,7 +4,7 @@ public class Adder {
 	
 	double sum;
 	
-	public Adder(int first, double second) {
+	public Adder(double first, double second) {
 		sum = first + second;
 	}
